Refactored daily notes and navigation arguments
This commit introduces significant changes to how daily notes are handled and how navigation arguments are passed, along with the introduction of a KSP processor for generating NavType maps.

**Daily Notes Refactor:**
- Removed `CreateDailyNotesFolderUseCase`, `GetDailyNotesFolderUseCase`, and `SetDailyNotesFolderNameUseCase` as daily notes are no longer stored in a dedicated, user-visible folder.
- Daily notes are now identified by a special `parentId` (-1) in `LocalPrefsDatasource`.
- `CreateDailyNoteUseCase` and `GetDailyNoteUseCase` now use this virtual folder ID.
- The `EditorViewModel` now checks if a note is a daily note (`isDailyNote`) and adjusts behavior accordingly (e.g., creating a new daily note if `arg` is `Route.Editor.Arg.NewDailyNote`).
- The editor screen's top bar title now reflects if it's a "Daily Note" or a regular "Note".
- A bottom bar with navigation for previous/next daily notes is now shown in the editor for daily notes.

**Navigation Argument Handling:**
- Introduced a `Route.Editor.Arg` sealed interface to handle different ways of opening the editor:
    - `NewDailyNote`: For creating a new daily note.
    - `NewNote(parentId: Int?)`: For creating a new regular note, optionally under a parent.
    - `Note(id: Int)`: For opening an existing note.
- `HomeViewModel` now uses these specific argument types when navigating to the editor.
- `EditorViewModel` receives `Route.Editor.Arg` to determine its initial state and actions (e.g., creating a new note vs. loading an existing one).
- Bottom bar navigation in `MainActivity` and `BottomBar.kt` updated to use the new `Route.Editor.Arg` types.

**KSP `GenerateTypeMapProcessor`:**
- Added a new module `typemap-processor` which includes a KSP (Kotlin Symbol Processing) processor.
- This processor generates `NavType` maps for classes annotated with `@GenerateTypeMap`. This simplifies passing complex objects (like `Route.Editor.Arg`) as navigation arguments by automatically handling serialization/deserialization.
- Integrated this processor into the `app` module.
- `Route.Editor.Arg` is annotated with `@GenerateTypeMap` to leverage this.

**Database & Data Model Changes:**
- `TagEntity` and `Tag` domain model now include an `automatic` boolean field (schema version updated).
- `NoteEntityParentIdPartial` now includes `updatedAt`.
- `LocalNoteDatasource` now has an `assignTags` method and `NoteDao` has corresponding methods to update note-tag cross-references.
- Removed `dailyNotesFolderName` and `dailyNotesFolderId` from `Preferences.kt` and `PreferencesDefaults.kt`.

**Other Changes:**
- Updated `EditorViewModel` to handle `parentId` when creating new notes.
- Small refactor in `LocalNoteDatasourceImpl` for `move` and added `assignTags` implementation.
- File renames for partial entity classes in `local-datasource`.
- Dependency updates in `gradle/libs.versions.toml` for KSP and KotlinPoet.

+++ local-datasource/src/main/java/com/pointlessapps/granite/local/datasource/note/entity/NoteTagCrossRef.kt
@@ -22,7 +22,7 @@ import androidx.room.ForeignKey
 )
 data class NoteTagCrossRef(
     @ColumnInfo(name = "note_id", index = true)
-    val noteId: Long,
+    val noteId: Int,
     @ColumnInfo(name = "tag_id", index = true)
-    val tagId: Long,
+    val tagId: Int,
 )
